feat: simplify and speed up selection of layer downsampling
